Refactor feedback form into report issue page
